backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
tensorflow>=2.16.0
Pillow>=10.0.0
numpy>=1.24.3